refactor: made the login card as an individual component
